lab/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-CORS==3.0.10
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Pillow==9.5.0
PyJWT==2.3.0
